# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LegalGuard AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Upload and create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "404":
          description: Not found
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contracts/{id}/analysis:
    get:
      operationId: getContractAnalysis
      tags: [contracts]
      summary: Get AI analysis for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractAnalysis"
...
  /receipts/{id}:
    get:
      operationId: getReceipt
      tags: [receipts]
      summary: Get a specific receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CryptoReceipt"
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: inviteTeamMember
      tags: [team]
      summary: Invite a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInvite"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    delete:
      operationId: removeTeamMember
      tags: [team]
      summary: Remove a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        client:
          type: string
        status:
          type: string
        riskScore:
          type: number
        riskLevel:
          type: string
        uploadedAt:
          type: string
        analyzedAt:
          type: ["string", "null"]
        fileType:
          type: string
        fileSize:
          type: string
        assignedTo:
          type: ["string", "null"]
      required: [id, name, client, status, riskScore, riskLevel, uploadedAt, fileType, fileSize]
    ContractDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        client:
          type: string
        status:
          type: string
        riskScore:
          type: number
        riskLevel:
          type: string
        uploadedAt:
          type: string
        analyzedAt:
          type: ["string", "null"]
        fileType:
          type: string
        fileSize:
          type: string
        assignedTo:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        clauses:
          type: array
          items:
            $ref: "#/components/schemas/Clause"
        obligations:
          type: array
          items:
            type: string
        deadlines:
          type: array
          items:
            type: string
      required: [id, name, client, status, riskScore, riskLevel, uploadedAt, fileType, fileSize, clauses, obligations, deadlines]
    Clause:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        riskLevel:
          type: string
        isRisky:
          type: boolean
        aiSuggestion:
          type: ["string", "null"]
      required: [id, title, content, riskLevel, isRisky]
    ContractAnalysis:
      type: object
      properties:
        contractId:
          type: integer
        overallRiskScore:
          type: number
        riskLevel:
          type: string
        executiveSummary:
          type: string
        riskyClauses:
          type: array
          items:
            $ref: "#/components/schemas/Clause"
        missingClauses:
          type: array
          items:
            type: string
        aiSuggestions:
          type: array
          items:
            type: string
        confidenceScore:
          type: number
        sensitiveActions:
          type: array
          items:
            type: string
      required: [contractId, overallRiskScore, riskLevel, executiveSummary, riskyClauses, missingClauses, aiSuggestions, confidenceScore, sensitiveActions]
    ContractInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        client:
          type: string
          minLength: 1
        fileType:
          type: string
        fileSize:
          type: string
        content:
          type: ["string", "null"]
      required: [name, client, fileType, fileSize]
    ContractUpdate:
      type: object
      properties:
        name:
          type: string
        client:
          type: string
        status:
          type: string
        assignedTo:
          type: ["string", "null"]
...
    CryptoReceipt:
      type: object
      properties:
        id:
          type: integer
        receiptId:
          type: string
        transactionHash:
          type: string
        documentName:
          type: string
        requestedBy:
          type: string
        approvedBy:
          type: string
        timestamp:
          type: string
        signature:
          type: string
        verificationStatus:
          type: string
        approvalId:
          type: ["integer", "null"]
      required: [id, receiptId, transactionHash, documentName, requestedBy, approvedBy, timestamp, signature, verificationStatus]
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        createdAt:
          type: string
        isRead:
          type: boolean
        cta:
          type: ["string", "null"]
        ctaLink:
          type: ["string", "null"]
      required: [id, title, description, category, status, createdAt, isRead]
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        joinedAt:
          type: string
        lastActive:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        actionsCount:
          type: integer
        approvalsCount:
          type: integer
      required: [id, name, email, role, status, joinedAt, actionsCount, approvalsCount]
    TeamMemberInvite:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
      required: [name, email, role]
...