# lib/api-zod/src/generated/types/clause.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LegalGuard AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Clause {
  id: number;
  title: string;
  content: string;
  riskLevel: string;
  isRisky: boolean;
  /** @nullable */
  aiSuggestion?: string | null;
}