# lib/api-zod/src/generated/types/contractDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LegalGuard AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Clause } from './clause';

export interface ContractDetail {
  id: number;
  name: string;
  client: string;
  status: string;
  riskScore: number;
  riskLevel: string;
  uploadedAt: string;
  /** @nullable */
  analyzedAt?: string | null;
  fileType: string;
  fileSize: string;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  summary?: string | null;
  clauses: Clause[];
  obligations: string[];
  deadlines: string[];
}