# lib/api-zod/src/generated/types/contractUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LegalGuard AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContractUpdate {
  name?: string;
  client?: string;
  status?: string;
  /** @nullable */
  assignedTo?: string | null;
}